issBackend/requirements.txt

djangorestframework==3.6.4
django-cors-headers==2.1.0
